+++ parler/cache.py
@@ -129,6 +129,9 @@ def _cache_translation(translation, timeout=DEFAULT_TIMEOUT):
     if not appsettings.PARLER_ENABLE_CACHING:
         return
 
+    if translation.master_id is None:
+        raise ValueError("Can't cache unsaved translation")
+
     # Cache a translation object.
     # For internal usage, object parameters are not suited for outside usage.
     fields = translation.get_translated_fields()