+++ CHANGES.rst
@@ -7,7 +7,7 @@ Changes in version 1.2 (in git)
 * **NOTE:** the ``TranslationDoesNotExist`` exception inherits from ``ObjectDoesNotExist`` now.
   Check your exception handlers when upgrading.
 * Added ``TranslatableAdmin.get_translation_objects()`` API.
-* Added ``TranslatableModel.add_translation()`` API.
+* Added ``TranslatableModel.create_translation()`` API.
 
 
 Changes in version 1.1.1 (2014-10-14)